Create role-menubar-menuitem-menuitemradio-menuitemcheckbox-cyns.html
Related to issue #26.
This is a static test case of correctly nested menu and menu items.  It has no script and does not do anything.  It tests whether the browser/os/at combination can read the role and state of the objects. the objects look like plain text in the visual rendering and are not functional.

It includes role=menubar, role=menuitem, role=menuitemradio, role=menuitemcheckbox, aria-haspopup and aria-checked.